fix compiler error
